Fixed audio source change edit

+++ app/Services/Stream/StreamsStore.php
@@ -2,6 +2,7 @@
 
 namespace App\Services\Stream;
 
+use App\ManualPresentation;
 use App\Stream;
 use App\StreamResolution;
 use Illuminate\Database\Eloquent\Model;
@@ -14,44 +15,48 @@ class StreamsStore extends Model
     {
         $this->streams = $request->input('package.sources');
         $this->video = $video;
-    }
+     }
 
     public function streams()
     {
-        if(count($this->streams)>0) {
-            foreach ($this->streams as $key => $source) {
-                if ($source) {
-                    $stream = Stream::updateOrCreate([
-                        'video_id' => $this->video->id, 'name' => $key ?? ''],
-                        [
-                        'poster' => $source['poster'] ?? '',
-                        'audio' => $source['playAudio'] ?? false
-                    ]);
+        //Check if presentation has been edited
+        if(!$presentation = ManualPresentation::where('pkg_id', $this->video->id)->where(function($query) {
+            $query->where('status', 'sent')->orWhere('status', 'completed');})->latest()->first()) {
+            if(count($this->streams)>0) {
+                foreach ($this->streams as $key => $source) {
+                    if ($source) {
+                        $stream = Stream::updateOrCreate([
+                            'video_id' => $this->video->id, 'name' => $key ?? ''],
+                            [
+                                'poster' => $source['poster'] ?? '',
+                                'audio' => $source['playAudio'] ?? false
+                            ]);
 
-                    if($source['video'] ?? false) {
-                        foreach ($source['video'] as $resolution => $url) {
-                            $streamresolution = StreamResolution::updateOrCreate(
-                                [
-                                'stream_id' => $stream->id,
-                                'resolution' => $resolution
-                                ],
-                                [
-                                'filename' => $url
-                                ]
-                            );
+                        if($source['video'] ?? false) {
+                            foreach ($source['video'] as $resolution => $url) {
+                                $streamresolution = StreamResolution::updateOrCreate(
+                                    [
+                                        'stream_id' => $stream->id,
+                                        'resolution' => $resolution
+                                    ],
+                                    [
+                                        'filename' => $url
+                                    ]
+                                );
+                            }
                         }
-                    }
-                } //end check
-            } // end foreach
+                    } //end check
+                } // end foreach
 
-        } else {
-            //Remove any old associations
-            $streams = Stream::where('video_id', $this->video->id)->get();
-            foreach ($streams as $stream) {
-                StreamResolution::where('stream_id', $stream->id)->delete();
-                $stream->delete();
+            } else {
+                //Remove any old associations
+                $streams = Stream::where('video_id', $this->video->id)->get();
+                foreach ($streams as $stream) {
+                    StreamResolution::where('stream_id', $stream->id)->delete();
+                    $stream->delete();
+                }
             }
-        }
+        } //end presentation edit
     }
 
 }
